requirements.txt
torch>=1.12.0
torchvision>=0.13.0
albumentations
pydicom
numpy
matplotlib
tqdm
thop
Pillow
pandas
seaborn
scikit-image